% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{create_ip_group}
\alias{create_ip_group}
\title{Creates an IP access control group}
\usage{
create_ip_group(
  GroupName = "string",
  GroupDesc = "string",
  UserRules = list(
    list(
      ipRule = "string",
      ruleDesc = "string"
    )
  )
)
}
\arguments{
\item{GroupName}{[required] The name of the group.}

\item{GroupDesc}{The description of the group.}

\item{UserRules}{The rules to add to the group.}
}
\description{
Creates an IP access control group.
}
\details{
An IP access control group provides you with the ability to control the IP addresses from which users are allowed to access their WorkSpaces. To specify the CIDR address ranges, add rules to your IP access control group and then associate the group with your directory. You can add rules when you create the group or at any time using AuthorizeIpRules.

There is a default IP access control group associated with your directory. If you don't associate an IP access control group with your directory, the default group is used. The default group includes a default rule that allows users to access their WorkSpaces from anywhere. You cannot modify the default IP access control group for your directory.
}
\examples{

}
