% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{delete_ip_group}
\alias{delete_ip_group}
\title{Deletes the specified IP access control group}
\usage{
delete_ip_group(GroupId)
}
\arguments{
\item{GroupId}{[required] The identifier of the IP access control group.}
}
\description{
Deletes the specified IP access control group.
}
\details{
You cannot delete an IP access control group that is associated with a directory.
}
\section{Accepted Parameters}{
\preformatted{delete_ip_group(
  GroupId = "string"
)
}
}

