% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{delete_tags}
\alias{delete_tags}
\title{Deletes the specified tags from the specified WorkSpace}
\usage{
delete_tags(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{The ID of the WorkSpace. To find this ID, use DescribeWorkspaces.}

\item{TagKeys}{The tag keys.}
}
\description{
Deletes the specified tags from the specified WorkSpace.
}
\examples{

}
