% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{delete_workspace_image}
\alias{delete_workspace_image}
\title{Deletes the specified image from your account}
\usage{
delete_workspace_image(ImageId)
}
\arguments{
\item{ImageId}{[required] The identifier of the image.}
}
\description{
Deletes the specified image from your account. To delete an image, you must first delete any bundles that are associated with the image.
}
\section{Accepted Parameters}{
\preformatted{delete_workspace_image(
  ImageId = "string"
)
}
}

