% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_account}
\alias{describe_account}
\title{Retrieves a list that describes the configuration of bring your own license (BYOL) for the specified account}
\usage{
describe_account()
}
\description{
Retrieves a list that describes the configuration of bring your own license (BYOL) for the specified account.
}
\examples{

}
