% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_account_modifications}
\alias{describe_account_modifications}
\title{Retrieves a list that describes modifications to the configuration of bring your own license (BYOL) for the specified account}
\usage{
describe_account_modifications(
  NextToken = "string"
)
}
\arguments{
\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated, provide this token to receive the next set of results.}
}
\description{
Retrieves a list that describes modifications to the configuration of bring your own license (BYOL) for the specified account.
}
\examples{

}
