% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_client_properties}
\alias{describe_client_properties}
\title{Retrieves a list that describes one or more specified Amazon WorkSpaces clients}
\usage{
describe_client_properties(
  ResourceIds = list(
    "string"
  )
)
}
\arguments{
\item{ResourceIds}{[required] The resource identifiers, in the form of directory IDs.}
}
\description{
Retrieves a list that describes one or more specified Amazon WorkSpaces clients.
}
\examples{

}
