% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_ip_groups}
\alias{describe_ip_groups}
\title{Describes one or more of your IP access control groups}
\usage{
describe_ip_groups(GroupIds = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{GroupIds}{The IDs of one or more IP access control groups.}

\item{NextToken}{The token for the next set of results. (You received this token from a previous call.)}

\item{MaxResults}{The maximum number of items to return.}
}
\description{
Describes one or more of your IP access control groups.
}
\section{Accepted Parameters}{
\preformatted{describe_ip_groups(
  GroupIds = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

