% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_tags}
\alias{describe_tags}
\title{Describes the specified tags for the specified WorkSpace}
\usage{
describe_tags(ResourceId)
}
\arguments{
\item{ResourceId}{The ID of the WorkSpace. To find this ID, use DescribeWorkspaces.}
}
\description{
Describes the specified tags for the specified WorkSpace.
}
\examples{

}
