% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_workspace_directories}
\alias{describe_workspace_directories}
\title{Describes the available AWS Directory Service directories that are registered with Amazon WorkSpaces}
\usage{
describe_workspace_directories(DirectoryIds = NULL, NextToken = NULL)
}
\arguments{
\item{DirectoryIds}{The identifiers of the directories. If the value is null, all directories are retrieved.}

\item{NextToken}{The token for the next set of results. (You received this token from a previous call.)}
}
\description{
Describes the available AWS Directory Service directories that are registered with Amazon WorkSpaces.
}
\examples{

}
