% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_workspace_images}
\alias{describe_workspace_images}
\title{Retrieves a list that describes one or more specified images, if the image identifiers are provided}
\usage{
describe_workspace_images(ImageIds = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{ImageIds}{The identifier of the image.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated, provide this token to receive the next set of results.}

\item{MaxResults}{The maximum number of items to return.}
}
\description{
Retrieves a list that describes one or more specified images, if the image identifiers are provided. Otherwise, all images in the account are described.
}
\section{Accepted Parameters}{
\preformatted{describe_workspace_images(
  ImageIds = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

