% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_workspaces}
\alias{describe_workspaces}
\title{Describes the specified WorkSpaces}
\usage{
describe_workspaces(WorkspaceIds = NULL, DirectoryId = NULL,
  UserName = NULL, BundleId = NULL, Limit = NULL, NextToken = NULL)
}
\arguments{
\item{WorkspaceIds}{The IDs of the WorkSpaces. This parameter cannot be combined with any other filter.

Because the CreateWorkspaces operation is asynchronous, the identifier it returns is not immediately available. If you immediately call DescribeWorkspaces with this identifier, no information is returned.}

\item{DirectoryId}{The ID of the directory. In addition, you can optionally specify a specific directory user (see \code{UserName}). This parameter cannot be combined with any other filter.}

\item{UserName}{The name of the directory user. You must specify this parameter with \code{DirectoryId}.}

\item{BundleId}{The ID of the bundle. All WorkSpaces that are created from this bundle are retrieved. This parameter cannot be combined with any other filter.}

\item{Limit}{The maximum number of items to return.}

\item{NextToken}{The token for the next set of results. (You received this token from a previous call.)}
}
\description{
Describes the specified WorkSpaces.
}
\details{
You can filter the results using bundle ID, directory ID, or owner, but you can specify only one filter at a time.
}
\section{Accepted Parameters}{
\preformatted{describe_workspaces(
  WorkspaceIds = list(
    "string"
  ),
  DirectoryId = "string",
  UserName = "string",
  BundleId = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

