% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{modify_client_properties}
\alias{modify_client_properties}
\title{Modifies the properties of the specified Amazon WorkSpaces client}
\usage{
modify_client_properties(
  ResourceId = "string",
  ClientProperties = list(
    ReconnectEnabled = "ENABLED"|"DISABLED"
  )
)
}
\arguments{
\item{ResourceId}{[required] The resource identifiers, in the form of directory IDs.}

\item{ClientProperties}{Information about the Amazon WorkSpaces client.}
}
\description{
Modifies the properties of the specified Amazon WorkSpaces client.
}
\examples{

}
