% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{modify_workspace_properties}
\alias{modify_workspace_properties}
\title{Modifies the specified WorkSpace properties}
\usage{
modify_workspace_properties(WorkspaceId, WorkspaceProperties)
}
\arguments{
\item{WorkspaceId}{The ID of the WorkSpace.}

\item{WorkspaceProperties}{The properties of the WorkSpace.}
}
\description{
Modifies the specified WorkSpace properties.
}
\examples{

}
