% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_package.R
\docType{package}
\name{paws.workspaces-package}
\alias{paws.workspaces}
\alias{paws.workspaces-package}
\title{paws.workspaces: Amazon WorkSpaces}
\description{
Amazon WorkSpaces enables you to provision virtual,
cloud-based Microsoft Windows and Amazon Linux desktops for your
users.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
