% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{reboot_workspaces}
\alias{reboot_workspaces}
\title{Reboots the specified WorkSpaces}
\usage{
reboot_workspaces(
  RebootWorkspaceRequests = list(
    list(
      WorkspaceId = "string"
    )
  )
)
}
\arguments{
\item{RebootWorkspaceRequests}{[required] The WorkSpaces to reboot. You can specify up to 25 WorkSpaces.}
}
\description{
Reboots the specified WorkSpaces.
}
\details{
You cannot reboot a WorkSpace unless its state is \code{AVAILABLE} or \code{UNHEALTHY}.

This operation is asynchronous and returns before the WorkSpaces have rebooted.
}
\examples{

}
