% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{rebuild_workspaces}
\alias{rebuild_workspaces}
\title{Rebuilds the specified WorkSpace}
\usage{
rebuild_workspaces(RebuildWorkspaceRequests)
}
\arguments{
\item{RebuildWorkspaceRequests}{The WorkSpace to rebuild. You can specify a single WorkSpace.}
}
\description{
Rebuilds the specified WorkSpace.
}
\details{
You cannot rebuild a WorkSpace unless its state is \code{AVAILABLE}, \code{ERROR}, or \code{UNHEALTHY}.

Rebuilding a WorkSpace is a potentially destructive action that can result in the loss of data. For more information, see \href{http://docs.aws.amazon.com/workspaces/latest/adminguide/reset-workspace.html}{Rebuild a WorkSpace}.

This operation is asynchronous and returns before the WorkSpaces have been completely rebuilt.
}
\examples{

}
