% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{revoke_ip_rules}
\alias{revoke_ip_rules}
\title{Removes one or more rules from the specified IP access control group}
\usage{
revoke_ip_rules(
  GroupId = "string",
  UserRules = list(
    "string"
  )
)
}
\arguments{
\item{GroupId}{[required] The identifier of the group.}

\item{UserRules}{[required] The rules to remove from the group.}
}
\description{
Removes one or more rules from the specified IP access control group.
}
\examples{

}
