% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{start_workspaces}
\alias{start_workspaces}
\title{Starts the specified WorkSpaces}
\usage{
start_workspaces(StartWorkspaceRequests)
}
\arguments{
\item{StartWorkspaceRequests}{The WorkSpaces to start. You can specify up to 25 WorkSpaces.}
}
\description{
Starts the specified WorkSpaces.
}
\details{
You cannot start a WorkSpace unless it has a running mode of \code{AutoStop} and a state of \code{STOPPED}.
}
\examples{

}
