% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{stop_workspaces}
\alias{stop_workspaces}
\title{Stops the specified WorkSpaces}
\usage{
stop_workspaces(StopWorkspaceRequests)
}
\arguments{
\item{StopWorkspaceRequests}{The WorkSpaces to stop. You can specify up to 25 WorkSpaces.}
}
\description{
Stops the specified WorkSpaces.
}
\details{
You cannot stop a WorkSpace unless it has a running mode of \code{AutoStop} and a state of \code{AVAILABLE}, \code{IMPAIRED}, \code{UNHEALTHY}, or \code{ERROR}.
}
\examples{

}
