% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{terminate_workspaces}
\alias{terminate_workspaces}
\title{Terminates the specified WorkSpaces}
\usage{
terminate_workspaces(TerminateWorkspaceRequests)
}
\arguments{
\item{TerminateWorkspaceRequests}{The WorkSpaces to terminate. You can specify up to 25 WorkSpaces.}
}
\description{
Terminates the specified WorkSpaces.
}
\details{
Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is destroyed. If you need to archive any user data, contact Amazon Web Services before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except \code{SUSPENDED}.

This operation is asynchronous and returns before the WorkSpaces have been completely terminated.
}
\examples{

}
