% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{authorize_ip_rules}
\alias{authorize_ip_rules}
\title{Adds one or more rules to the specified IP access control group}
\usage{
authorize_ip_rules(GroupId, UserRules)
}
\arguments{
\item{GroupId}{The ID of the group.}

\item{UserRules}{The rules to add to the group.}
}
\description{
Adds one or more rules to the specified IP access control group.
}
\details{
This action gives users permission to access their WorkSpaces from the CIDR address ranges specified in the rules.
}
\examples{

}
