% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_ip_groups}
\alias{describe_ip_groups}
\title{Describes one or more of your IP access control groups}
\usage{
describe_ip_groups(
  GroupIds = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{GroupIds}{The identifiers of one or more IP access control groups.}

\item{NextToken}{If you received a \code{NextToken} from a previous call that was paginated, provide this token to receive the next set of results.}

\item{MaxResults}{The maximum number of items to return.}
}
\description{
Describes one or more of your IP access control groups.
}
\examples{

}
