% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{describe_workspaces_connection_status}
\alias{describe_workspaces_connection_status}
\title{Describes the connection status of the specified WorkSpaces}
\usage{
describe_workspaces_connection_status(WorkspaceIds = NULL,
  NextToken = NULL)
}
\arguments{
\item{WorkspaceIds}{The identifiers of the WorkSpaces. You can specify up to 25 WorkSpaces.}

\item{NextToken}{The token for the next set of results. (You received this token from a previous call.)}
}
\description{
Describes the connection status of the specified WorkSpaces.
}
\section{Accepted Parameters}{
\preformatted{describe_workspaces_connection_status(
  WorkspaceIds = list(
    "string"
  ),
  NextToken = "string"
)
}
}

