% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{disassociate_ip_groups}
\alias{disassociate_ip_groups}
\title{Disassociates the specified IP access control group from the specified directory}
\usage{
disassociate_ip_groups(DirectoryId, GroupIds)
}
\arguments{
\item{DirectoryId}{The ID of the directory.}

\item{GroupIds}{The IDs of one or more IP access control groups.}
}
\description{
Disassociates the specified IP access control group from the specified directory.
}
\examples{

}
