% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workspaces_operations.R
\name{revoke_ip_rules}
\alias{revoke_ip_rules}
\title{Removes one or more rules from the specified IP access control group}
\usage{
revoke_ip_rules(GroupId, UserRules)
}
\arguments{
\item{GroupId}{The ID of the group.}

\item{UserRules}{The rules to remove from the group.}
}
\description{
Removes one or more rules from the specified IP access control group.
}
\examples{

}
